add FixedBuffer_test

+++ src/base/AsyncLogging.cpp
@@ -93,7 +93,7 @@ void AsyncLogging::threadFunc()
     output.flush();
 }
 
-void AsyncLogging::append(const char* logline, int len)
+void AsyncLogging::append(const char* logline, size_t len)
 {
     MutexLockGuard lock(mutex_);
     if (currentBuffer_->avail() > len) {
